Company/Document Logo
Parts of the client are still using old company vs organization; a big part of this is fetching company logos to display.

We need to sync up the documents that correspond to a given potential parent, in this case organizations. As a result, created a listener that will do this.